+++ VersionSet.xml
@@ -1,0 +1,1 @@
+<version_set type="microService"></version_set>
