--- ZenPack/community/OceanStor/objects/objects.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0"?>
-<objects>
-</objects>
