--- ZenPack/community/OceanStor/objects/templates/Devices_Storage_Huawei_OceanStor_rrdTemplates_OceanStorArray.xml
@@ -1,203 +1,0 @@
-<?xml version="1.0"?>
-<objects>
-<!-- ('', 'zport', 'dmd', 'Devices', 'Storage', 'Huawei', 'OceanStor', 'rrdTemplates', 'OceanStorArray') -->
-<object id='/zport/dmd/Devices/Storage/Huawei/OceanStor/rrdTemplates/OceanStorArray' module='Products.ZenModel.RRDTemplate' class='RRDTemplate'>
-<property id='zendoc' type='string'>
-OceanStor array monitoring.
-</property>
-<property type="text" id="description" mode="w" >
-OceanStor array monitoring.
-</property>
-<property type="string" id="targetPythonClass" mode="w" >
-ZenPacks.community.OceanStor.OceanStorArray
-</property>
-<tomanycont id='datasources'>
-<object id='array' module='ZenPacks.zenoss.PythonCollector.datasources.PythonDataSource' class='PythonDataSource'>
-<property select_variable="sourcetypes" type="selection" id="sourcetype" mode="w" >
-Python
-</property>
-<property type="boolean" id="enabled" mode="w" >
-True
-</property>
-<property type="int" id="severity" mode="w" >
-3
-</property>
-<property type="string" id="cycletime" mode="w" >
-${here/zCommandCollectionInterval}
-</property>
-<property type="string" id="plugin_classname" mode="w" >
-ZenPacks.community.OceanStor.dsplugins.Array
-</property>
-<tomanycont id='datapoints'>
-<object id='STORAGE_POOLS_CAPACITY_USAGE' module='Products.ZenModel.RRDDataPoint' class='RRDDataPoint'>
-<property select_variable="rrdtypes" type="selection" id="rrdtype" mode="w" >
-GAUGE
-</property>
-<property type="boolean" id="isrow" mode="w" >
-True
-</property>
-</object>
-<object id='SYSTEM_CAPACITY_USAGE' module='Products.ZenModel.RRDDataPoint' class='RRDDataPoint'>
-<property select_variable="rrdtypes" type="selection" id="rrdtype" mode="w" >
-GAUGE
-</property>
-<property type="boolean" id="isrow" mode="w" >
-True
-</property>
-</object>
-</tomanycont>
-</object>
-</tomanycont>
-<tomanycont id='thresholds'>
-<object id='storage_pools_capacity_usage' module='Products.ZenModel.MinMaxThreshold' class='MinMaxThreshold'>
-<property type="string" id="eventClass" mode="w" >
-/Status
-</property>
-<property type="int" id="severity" mode="w" >
-3
-</property>
-<property label="DataPoints" type="lines" id="dsnames" mode="w" >
-['array_STORAGE_POOLS_CAPACITY_USAGE']
-</property>
-<property label="Enabled" type="boolean" id="enabled" mode="w" >
-True
-</property>
-<property label="Is Projection" type="boolean" id="isProjection" mode="w" >
-False
-</property>
-<property type="string" id="maxval" mode="w" >
-80
-</property>
-<property type="int" id="escalateCount" mode="w" >
-0
-</property>
-</object>
-<object id='system_capacity_usage' module='Products.ZenModel.MinMaxThreshold' class='MinMaxThreshold'>
-<property type="string" id="eventClass" mode="w" >
-/Status
-</property>
-<property type="int" id="severity" mode="w" >
-3
-</property>
-<property label="DataPoints" type="lines" id="dsnames" mode="w" >
-['array_SYSTEM_CAPACITY_USAGE']
-</property>
-<property label="Enabled" type="boolean" id="enabled" mode="w" >
-True
-</property>
-<property label="Is Projection" type="boolean" id="isProjection" mode="w" >
-False
-</property>
-<property type="string" id="maxval" mode="w" >
-80
-</property>
-<property type="int" id="escalateCount" mode="w" >
-0
-</property>
-</object>
-</tomanycont>
-<tomanycont id='graphDefs'>
-<object id='Capacity Usage' module='Products.ZenModel.GraphDefinition' class='GraphDefinition'>
-<property type="int" id="height" mode="w" >
-500
-</property>
-<property type="int" id="width" mode="w" >
-500
-</property>
-<property type="string" id="units" mode="w" >
-%
-</property>
-<property type="boolean" id="log" mode="w" >
-False
-</property>
-<property type="boolean" id="base" mode="w" >
-False
-</property>
-<property type="boolean" id="autoscale" mode="w" >
-None
-</property>
-<property type="int" id="miny" mode="w" >
-0
-</property>
-<property type="int" id="ceiling" mode="w" >
-None
-</property>
-<property type="int" id="maxy" mode="w" >
-100
-</property>
-<property type="boolean" id="hasSummary" mode="w" >
-True
-</property>
-<property type="long" id="sequence" mode="w" >
-0
-</property>
-<tomanycont id='graphPoints'>
-<object id='Storage Pools' module='Products.ZenModel.DataPointGraphPoint' class='DataPointGraphPoint'>
-<property type="long" id="sequence" mode="w" >
-1
-</property>
-<property select_variable="lineTypes" type="selection" id="lineType" mode="w" >
-LINE
-</property>
-<property type="long" id="lineWidth" mode="w" >
-1
-</property>
-<property type="boolean" id="stacked" mode="w" >
-False
-</property>
-<property type="string" id="format" mode="w" >
-%5.0lf%s
-</property>
-<property type="string" id="legend" mode="w" >
-${graphPoint/id}
-</property>
-<property type="long" id="limit" mode="w" >
--1
-</property>
-<property type="string" id="dpName" mode="w" >
-array_STORAGE_POOLS_CAPACITY_USAGE
-</property>
-<property type="string" id="cFunc" mode="w" >
-AVERAGE
-</property>
-<property type="boolean" id="skipCalc" mode="w" >
-False
-</property>
-</object>
-<object id='System' module='Products.ZenModel.DataPointGraphPoint' class='DataPointGraphPoint'>
-<property type="long" id="sequence" mode="w" >
-0
-</property>
-<property select_variable="lineTypes" type="selection" id="lineType" mode="w" >
-LINE
-</property>
-<property type="long" id="lineWidth" mode="w" >
-1
-</property>
-<property type="boolean" id="stacked" mode="w" >
-False
-</property>
-<property type="string" id="format" mode="w" >
-%5.0lf%s
-</property>
-<property type="string" id="legend" mode="w" >
-${graphPoint/id}
-</property>
-<property type="long" id="limit" mode="w" >
--1
-</property>
-<property type="string" id="dpName" mode="w" >
-array_SYSTEM_CAPACITY_USAGE
-</property>
-<property type="string" id="cFunc" mode="w" >
-AVERAGE
-</property>
-<property type="boolean" id="skipCalc" mode="w" >
-False
-</property>
-</object>
-</tomanycont>
-</object>
-</tomanycont>
-</object>
-</objects>
