--- ZenPack/community/OceanStor/objects/templates/Devices_Storage_Huawei_OceanStor_rrdTemplates_OceanStorEnclosure.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0"?>
-<objects>
-<!-- ('', 'zport', 'dmd', 'Devices', 'Storage', 'Huawei', 'OceanStor', 'rrdTemplates', 'OceanStorEnclosure') -->
-<object id='/zport/dmd/Devices/Storage/Huawei/OceanStor/rrdTemplates/OceanStorEnclosure' module='Products.ZenModel.RRDTemplate' class='RRDTemplate'>
-<property id='zendoc' type='string'>
-OceanStor enclosure monitoring.
-</property>
-<property type="text" id="description" mode="w" >
-OceanStor enclosure monitoring.
-</property>
-<property type="string" id="targetPythonClass" mode="w" >
-ZenPacks.community.OceanStor.OceanStorEnclosure
-</property>
-<tomanycont id='datasources'>
-<object id='enclosure' module='ZenPacks.zenoss.PythonCollector.datasources.PythonDataSource' class='PythonDataSource'>
-<property select_variable="sourcetypes" type="selection" id="sourcetype" mode="w" >
-Python
-</property>
-<property type="boolean" id="enabled" mode="w" >
-True
-</property>
-<property type="int" id="severity" mode="w" >
-3
-</property>
-<property type="string" id="cycletime" mode="w" >
-${here/zCommandCollectionInterval}
-</property>
-<property type="string" id="plugin_classname" mode="w" >
-ZenPacks.community.OceanStor.dsplugins.Enclosure
-</property>
-<tomanycont id='datapoints'>
-<object id='TEMPERATURE' module='Products.ZenModel.RRDDataPoint' class='RRDDataPoint'>
-<property select_variable="rrdtypes" type="selection" id="rrdtype" mode="w" >
-GAUGE
-</property>
-<property type="boolean" id="isrow" mode="w" >
-True
-</property>
-</object>
-</tomanycont>
-</object>
-</tomanycont>
-<tomanycont id='thresholds'>
-<object id='temperature' module='Products.ZenModel.MinMaxThreshold' class='MinMaxThreshold'>
-<property type="string" id="eventClass" mode="w" >
-/Status
-</property>
-<property type="int" id="severity" mode="w" >
-3
-</property>
-<property label="DataPoints" type="lines" id="dsnames" mode="w" >
-['enclosure_TEMPERATURE']
-</property>
-<property label="Enabled" type="boolean" id="enabled" mode="w" >
-True
-</property>
-<property label="Is Projection" type="boolean" id="isProjection" mode="w" >
-False
-</property>
-<property type="string" id="maxval" mode="w" >
-48
-</property>
-<property type="int" id="escalateCount" mode="w" >
-0
-</property>
-</object>
-</tomanycont>
-</object>
-</objects>
