--- ZenPack/community/OceanStor/objects/templates/Devices_Storage_Huawei_OceanStor_rrdTemplates_OceanStorEvents.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0"?>
-<objects>
-<!-- ('', 'zport', 'dmd', 'Devices', 'Storage', 'Huawei', 'OceanStor', 'rrdTemplates', 'OceanStorEvents') -->
-<object id='/zport/dmd/Devices/Storage/Huawei/OceanStor/rrdTemplates/OceanStorEvents' module='Products.ZenModel.RRDTemplate' class='RRDTemplate'>
-<property type="string" id="targetPythonClass" mode="w" >
-ZenPacks.community.OceanStor.OceanStorDevice
-</property>
-<tomanycont id='datasources'>
-<object id='alarms' module='ZenPacks.zenoss.PythonCollector.datasources.PythonDataSource' class='PythonDataSource'>
-<property select_variable="sourcetypes" type="selection" id="sourcetype" mode="w" >
-Python
-</property>
-<property type="boolean" id="enabled" mode="w" >
-True
-</property>
-<property type="int" id="severity" mode="w" >
-3
-</property>
-<property type="string" id="cycletime" mode="w" >
-${here/zCommandCollectionInterval}
-</property>
-<property type="string" id="plugin_classname" mode="w" >
-ZenPacks.community.OceanStor.dsplugins.Alarms
-</property>
-</object>
-</tomanycont>
-</object>
-</objects>
