+++ ZenPacks.community.OceanStor/MANIFEST.in
@@ -1,0 +1,1 @@
+graft ZenPacks
