+++ ZenPacks.community.OceanStor/ZenPacks/community/OceanStor/lib/utils.py
