+++ ZenPacks.community.OceanStor/ZenPacks/community/OceanStor/modeler/plugins/community/api/OceanStorArray.py
@@ -1,0 +1,28 @@
+"""Models Huawei OceanStor Storage using the Storage REST API."""
+
+from ZenPacks.community.OceanStor.lib.model import OceanStorModel, OceanStorBase
+
+
+class _OceanStorArray(OceanStorModel):
+    relname = 'oceanStorArrays'
+    modname = 'ZenPacks.community.OceanStor.OceanStorArray'
+
+    FIELD_PARSER = {
+        'ID': (OceanStorModel._parse_id, 'id'),
+        'NAME': (OceanStorModel._parse_str, 'title'),
+        'HEALTHSTATUS': (OceanStorModel._parse_health_status, None),
+        'RUNNINGSTATUS': (OceanStorModel._parse_running_status, None),
+        'productModeString': (OceanStorModel._parse_product_model, None),
+        'pointRelease': (OceanStorModel._parse_product_version, 'productVersion'),
+        'TOTALCAPACITY': (OceanStorModel._parse_capacity, None),
+        'USEDCAPACITY': (OceanStorModel._parse_capacity, None),
+        'STORAGEPOOLCAPACITY': (OceanStorModel._parse_capacity, None),
+        'STORAGEPOOLFREECAPACITY': (OceanStorModel._parse_capacity, None),
+    }
+
+    def get_data(self, __):
+        return [self.cli.get_system()]
+
+
+class OceanStorArray(OceanStorBase):
+    DEVICE_MODELERS = {'OceanStorArray': _OceanStorArray}
