+++ app_define.json
@@ -1,0 +1,1 @@
+{"fileVersion":"1","name":"ESDK_Enterprise_Storage_ZenPack","description":"","version":"1.0.0","type":"microService","processes":{"ESDK_Enterprise_Storage_ZenPack":{"subscribes":[]}}}
